move detectAnchor function to scripts.js
