+++ scripts/lib-franklin.js
@@ -182,34 +182,6 @@ export function decorateIcons(element = document) {
 }
 
 /**
- * Detect anchor
- */
-export function detectAnchor(block) {
-  const activeHash = window.location.hash;
-  if (!activeHash) return;
-
-  const id = activeHash.substring(1, activeHash.length).toLocaleLowerCase();
-  const el = block.querySelector(`#${id}`);
-  if (el) {
-    const observer = new MutationObserver((mutationList) => {
-      mutationList.forEach((mutation) => {
-        if (mutation.type === 'attributes'
-          && mutation.attributeName === 'data-block-status'
-          && block.attributes.getNamedItem('data-block-status').value === 'loaded') {
-          observer.disconnect();
-          setTimeout(() => {
-            window.dispatchEvent(new Event('hashchange'));
-          },
-          3500,
-          );
-        }
-      });
-    });
-    observer.observe(block, { attributes: true });
-  }
-}
-
-/**
  * Gets placeholders object
  * @param {string} prefix
  */
